update: 更新文件 net_ethernet_type.h

+++ network/netmanager_ext/net_ethernet/net_ethernet_type.h
@@ -101,7 +101,7 @@ typedef struct Ethernet_NetAddr {
      * Port number
      * @since 26.0.0
      */
-    uint8_t port;
+    uint16_t port;
     /**
      * IP Address
      * @since 26.0.0
